move loss space
